chore: migrate tests to TS, rm mocks - rm unnecessary re-render tests since those basically tests if React works - rm mocks and use real components. Mocks were hiding bugs before - add `generateIntlContext` to src since it's a very useful test utils
