+++ src/index.ts
@@ -22,3 +22,4 @@ export {default as FormattedNumber} from './components/number';
 export {default as FormattedPlural} from './components/plural';
 export {default as FormattedMessage} from './components/message';
 export {default as FormattedHTMLMessage} from './components/html-message';
+export {generateIntlContext} from './test-utils';
