+++ test/functional/support/build.ts
@@ -1,4 +1,4 @@
-import * as ReactIntl from '../../../src/';
+import * as ReactIntl from '../../../src';
 
 export default function(buildPath) {
   describe('build', () => {
