+++ test/unit/components/html-message.tsx
@@ -1,12 +1,13 @@
 import * as React from 'react';
-import {generateIntlContext, makeMockContext, shallowDeep} from '../testUtils';
+import {mountFormattedComponentWithProvider} from '../testUtils';
 import FormattedHTMLMessage, {
   BaseFormattedHTMLMessage,
 } from '../../../src/components/html-message';
 import {BaseFormattedMessage} from '../../../src/components/message';
+import {generateIntlContext} from '../../../src/test-utils';
 
-const mockContext = makeMockContext(
-  require.resolve('../../../src/components/html-message')
+const mountWithProvider = mountFormattedComponentWithProvider(
+  FormattedHTMLMessage
 );
 
 describe('<FormattedHTMLMessage>', () => {
@@ -18,6 +19,7 @@ describe('<FormattedHTMLMessage>', () => {
     intl = generateIntlContext({
       locale: 'en',
       defaultLocale: 'en-US',
+      textComponent: 'span',
     });
   });
 
@@ -34,106 +36,92 @@ describe('<FormattedHTMLMessage>', () => {
   });
 
   it('renders a formatted HTML message in a <span>', () => {
-    const FormattedHTMLMessage = mockContext(intl);
     const descriptor = {
       id: 'hello',
       defaultMessage: 'Hello, <b>World</b>!',
     };
 
-    const rendered = shallowDeep(<FormattedHTMLMessage {...descriptor} />, 2);
+    const rendered = mountWithProvider(descriptor, intl).find('span');
 
-    expect(typeof rendered.type()).toBe('symbol');
     expect(rendered.prop('dangerouslySetInnerHTML')).toEqual({
       __html: intl.formatHTMLMessage(descriptor),
     });
   });
 
   it('accepts `values` prop', () => {
-    const FormattedHTMLMessage = mockContext(intl);
     const descriptor = {
       id: 'hello',
       defaultMessage: 'Hello, <b>{name}</b>!',
     };
     const values = {name: 'Eric'};
 
-    const rendered = shallowDeep(
-      <FormattedHTMLMessage {...descriptor} values={values} />,
-      2
+    const rendered = mountWithProvider({...descriptor, values}, intl).find(
+      'span'
     );
 
-    expect(rendered.prop('dangerouslySetInnerHTML').__html).toBe(
+    expect((rendered.prop('dangerouslySetInnerHTML') as any).__html).toBe(
       intl.formatHTMLMessage(descriptor, values)
     );
   });
 
   it('should HTML-escape `values`', () => {
-    const FormattedHTMLMessage = mockContext(intl);
     const descriptor = {
       id: 'hello',
       defaultMessage: 'Hello, <b>{name}</b>!',
     };
     const values = {name: '<i>Eric</i>'};
 
-    const rendered = shallowDeep(
-      <FormattedHTMLMessage {...descriptor} values={values} />,
-      2
+    const rendered = mountWithProvider({...descriptor, values}, intl).find(
+      'span'
     );
 
-    expect(rendered.prop('dangerouslySetInnerHTML').__html).toBe(
+    expect((rendered.prop('dangerouslySetInnerHTML') as any).__html).toBe(
       'Hello, <b>&lt;i&gt;Eric&lt;/i&gt;</b>!'
     );
   });
 
   it('accepts `tagName` prop', () => {
-    const FormattedHTMLMessage = mockContext(intl);
     const descriptor = {
       id: 'hello',
       defaultMessage: 'Hello, <b>World</b>!',
     };
     const tagName = 'p';
 
-    const rendered = shallowDeep(
-      <FormattedHTMLMessage {...descriptor} tagName={tagName} />,
-      2
+    const rendered = mountWithProvider({...descriptor, tagName}, intl).find(
+      tagName
     );
 
     expect(rendered.type()).toEqual(tagName);
   });
 
   it('supports function-as-child pattern', () => {
-    const FormattedHTMLMessage = mockContext(intl);
     const descriptor = {
       id: 'hello',
       defaultMessage: 'Hello, <b>World</b>!',
     };
 
     const spy = jest.fn().mockImplementation(() => <p>Jest</p>);
-    const rendered = shallowDeep(
-      <FormattedHTMLMessage {...descriptor}>{spy}</FormattedHTMLMessage>,
-      2
-    );
+    const rendered = mountWithProvider({...descriptor, children: spy}, intl);
 
     expect(spy).toHaveBeenCalledTimes(1);
     expect(spy.mock.calls[0]).toEqual([intl.formatHTMLMessage(descriptor)]);
 
-    expect(rendered.type()).toBe('p');
     expect(rendered.text()).toBe('Jest');
   });
 
   it('does not support rich-text message formatting', () => {
-    const FormattedHTMLMessage = mockContext(intl);
-    const rendered = shallowDeep(
-      <FormattedHTMLMessage
-        id="hello"
-        defaultMessage="Hello, <b>{name}</b>!"
-        values={{
+    const rendered = mountWithProvider(
+      {
+        id: 'hello',
+        defaultMessage: 'Hello, <b>{name}</b>!',
+        values: {
           name: <i>Eric</i>,
-        }}
-      />,
-      2
-    );
+        },
+      },
+      intl
+    ).find('span');
 
-    expect(rendered.prop('dangerouslySetInnerHTML').__html).toBe(
+    expect((rendered.prop('dangerouslySetInnerHTML') as any).__html).toBe(
       'Hello, <b>[object Object]</b>!'
     );
   });
